service2/requirements.txt
grpcio
grpcio-tools
sqlmodel
fastapi
uvicorn
protobuf